chore: add generate eval button to the connect modal
