+++ apps/web/src/app/(private)/projects/[projectId]/versions/[commitUuid]/documents/[documentUuid]/evaluations/dashboard/connect/_components/ConnectEvaluationModal/List/index.tsx
@@ -1,6 +1,17 @@
 import React from 'react'
 
-import { Input, SelectableCard, Skeleton, Text } from '@latitude-data/web-ui'
+import {
+  Input,
+  SelectableCard,
+  Skeleton,
+  Text,
+  useCurrentCommit,
+  useCurrentProject,
+} from '@latitude-data/web-ui'
+import { useCurrentDocument } from '$/app/providers/DocumentProvider'
+import { ROUTES } from '$/services/routes'
+import useCurrentWorkspace from '$/stores/currentWorkspace'
+import Link from 'next/link'
 
 interface EvaluationListProps {
   items: {
@@ -22,6 +33,11 @@ export function EvaluationList({
   searchTerm,
   onSearchChange,
 }: EvaluationListProps) {
+  const { project } = useCurrentProject()
+  const { commit } = useCurrentCommit()
+  const document = useCurrentDocument()
+  const workspace = useCurrentWorkspace()
+
   return (
     <div className='w-1/2 max-h-[520px] overflow-y-auto'>
       <Input
@@ -31,17 +47,47 @@ export function EvaluationList({
         onChange={(e) => onSearchChange(e.target.value)}
         className='w-full p-2 mb-4 border rounded-lg'
       />
-      <ul className='space-y-2'>
-        {items.map((item) => (
-          <SelectableCard
-            key={item.uuid}
-            title={item.name}
-            description={item.description}
-            selected={item.uuid === selectedItem}
-            onClick={() => onSelectItem(item.uuid)}
-          />
-        ))}
-      </ul>
+
+      <div className='space-y-6'>
+        {workspace.data.id === 1 && (
+          <div className='flex flex-col gap-2'>
+            <Text.H6M color='foregroundMuted'>Suggested</Text.H6M>
+            <Link
+              href={
+                ROUTES.projects
+                  .detail({ id: project.id })
+                  .commits.detail({ uuid: commit.uuid })
+                  .documents.detail({ uuid: document.documentUuid }).evaluations
+                  .dashboard.generate.root
+              }
+            >
+              <div className='flex flex-col gap-2 p-2 rounded-lg border hover:bg-secondary'>
+                <Text.H5M>AI generated eval</Text.H5M>
+                <Text.H6 color='foregroundMuted'>
+                  Our AI will craft an evaluation template just for this
+                  specific prompt. Try it out!
+                </Text.H6>
+              </div>
+            </Link>
+          </div>
+        )}
+        {items.length > 0 && (
+          <div className='flex flex-col gap-2'>
+            <Text.H6M color='foregroundMuted'>Templates</Text.H6M>
+            <ul className='space-y-2'>
+              {items.map((item) => (
+                <SelectableCard
+                  key={item.uuid}
+                  title={item.name}
+                  description={item.description}
+                  selected={item.uuid === selectedItem}
+                  onClick={() => onSelectItem(item.uuid)}
+                />
+              ))}
+            </ul>
+          </div>
+        )}
+      </div>
       {items.length === 0 && (
         <div className='text-center'>
           <Text.H6 color='foregroundMuted'>
